updating or adding when merging in the main trunk with the version in the branch nadir_1_2_2 !!

+++ src/SMESHGUI/SMESHGUI.h
@@ -150,7 +150,7 @@ public :
   void RemoveHypothesisOrAlgorithmOnMesh( SALOMEDS::SObject_ptr MorSM, SMESH::SMESH_Hypothesis_ptr anHyp ) ;
 
   void CreateLocalLength( QString TypeHypothesis, QString NameHypothesis, double Length );
-  void CreateNbSegments( QString TypeHypothesis, QString NameHypothesis, double nbSegments );
+  void CreateNbSegments( QString TypeHypothesis, QString NameHypothesis, int nbSegments );
   void CreateMaxElementArea( QString TypeHypothesis, QString NameHypothesis, double MaxArea );
   void CreateMaxElementVolume( QString TypeHypothesis, QString NameHypothesis, double MaxVolume );
